Rename schedule api file to be consistent with other api file namings
