+++ src/components/Patient/PatientDetailsTab/Appointments.tsx
@@ -20,7 +20,7 @@ import { PatientProps } from "@/components/Patient/PatientDetailsTab";
 
 import query from "@/Utils/request/query";
 import { formatDateTime, formatName } from "@/Utils/utils";
-import scheduleApis from "@/types/scheduling/scheduleApis";
+import scheduleApis from "@/types/scheduling/scheduleApi";
 
 export const Appointments = (props: PatientProps) => {
   const { patientData, facilityId, patientId } = props;
