+++ src/pages/Scheduling/components/CreateScheduleExceptionSheet.tsx
@@ -34,7 +34,7 @@ import {
 import mutate from "@/Utils/request/mutate";
 import { Time } from "@/Utils/types";
 import { dateQueryString } from "@/Utils/utils";
-import scheduleApis from "@/types/scheduling/scheduleApis";
+import scheduleApis from "@/types/scheduling/scheduleApi";
 
 interface Props {
   facilityId: string;
